refactor: domain 과 view 분리
domain 패키지 내부엔 핵심 비지니스 로직을 담당하는 클래스를 위치 -> Ladder, RandomGenerator
view 패키지 내부엔 UI 관련된 클래스를 위치 -> UI
