+++ src/main/java/kr/codesquad/domain/RandomGenerator.java
@@ -1,4 +1,4 @@
-package kr.codesquad;
+package kr.codesquad.domain;
 
 import java.util.ArrayList;
 import java.util.List;
