+++ src/main/java/kr/codesquad/view/UI.java
@@ -1,4 +1,6 @@
-package kr.codesquad;
+package kr.codesquad.view;
+
+import kr.codesquad.domain.Ladder;
 
 import java.util.List;
