+++ src/test/java/kr/codesquad/domain/LadderTest.java
@@ -1,6 +1,6 @@
-package kr.codesquad;
+package kr.codesquad.domain;
 
-import org.assertj.core.api.Assertions;
+import kr.codesquad.domain.Ladder;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
